Separate the edit and deleteOne functions
